Refactors
- Move getting of filter values into main script (to decouple filter getters)
- Remove getMoveString and implement getString properly on Move
