+++ src/js/filters.js
@@ -1,10 +1,9 @@
 /**
  * @param {Move[]} moves
+ * @param {Object} filters
  * @return {Move[]}
  */
-export function filterMoveList(moves) {
-	let filters = getFilters();
-
+export function filterMoveList(moves, filters) {
 	return moves.filter(function(move) {
 		let moveString = move.getString();
 		let includeMove = true;
@@ -55,43 +54,6 @@ export function filterMoveList(moves) {
 
 		return includeMove;
 	});
-}
-
-// impure - document
-function getFilters() {
-	let moveName = document.querySelector('#move-name-filter').value;
-	let moveString = document.querySelector('#move-string-filter').value;
-	let specialProperties = {
-		spin: document.querySelector('#move-property-spin-filter').checked,
-		track: document.querySelector('#move-property-track-filter').checked,
-		armor: document.querySelector('#move-property-armor-filter').checked,
-	};
-
-	let frameProperties = {
-		start: {
-			value: document.querySelector('#frame-property-start-filter').value,
-			comparison: document.querySelector('#frame-property-start-comparison-filter').value,
-		},
-		block: {
-			value: document.querySelector('#frame-property-block-filter').value,
-			comparison: document.querySelector('#frame-property-block-comparison-filter').value,
-		},
-		hit: {
-			value: document.querySelector('#frame-property-hit-filter').value,
-			comparison: document.querySelector('#frame-property-hit-comparison-filter').value,
-		}
-	};
-
-	frameProperties.start.value = parseInt(frameProperties.start.value);
-	frameProperties.block.value = parseInt(frameProperties.block.value);
-	frameProperties.hit.value   = parseInt(frameProperties.hit.value);
-
-	return {
-		moveName: moveName,
-		moveString: moveString,
-		specialProperties: specialProperties,
-		frameProperties: frameProperties,
-	};
 }
 
 function compare(value1, value2, operator) {
