+++ src/js/move.js
@@ -1,42 +1,3 @@
-import { isLetter } from './utils.js';
-
-// impure
-export function getMoveString(move, lang, ctrlsMap) {
-    let moveString = "";
-    let commands = move.command[lang].split(" ");
-
-    for (let c = 0; c < commands.length; c++) {
-        let command = commands[c];
-
-        if (/[a-z]/.test(command.toLowerCase())) {
-            moveString += command;
-        } else {
-            for (let m = 0; m < command.length; m++) {
-                let input = "";
-                try {
-                    if (isLetter(ctrlsMap[command.charAt(m)])) {
-                        if (
-                            ctrlsMap[command.charAt(m)] === ctrlsMap[command.charAt(m)].toLowerCase() ||
-                            ctrlsMap[command.charAt(m)] === "N"
-                        ) {
-                            input = ctrlsMap[command.charAt(m)].toLowerCase();
-                        } else {
-                            input = ctrlsMap[command.charAt(m)].toLowerCase();
-                        }
-                    } else if (!isNaN(ctrlsMap[command.charAt(m)].charAt(0))) {
-                        input = ctrlsMap[command.charAt(m)];
-                    }
-                } catch (exception) {
-                }
-
-                moveString += input;
-            }
-        }
-    }
-
-    return moveString;
-}
-
 export default class Move {
     constructor(moveData, language, controlsMap, hitsMap) {
         this._moveData = moveData;
@@ -79,7 +40,11 @@ export default class Move {
     }
 
     getString() {
-        return getMoveString(this._moveData, this._language, this._controlsMap);
+        return this.getCommands()
+            .map((command) => {
+                return command.getInputs()
+                    .map((input) => input.getSymbol());
+            }).join("");
     }
 
     hasThrow() {
